+++ files/src/sdk/sdk.ts
@@ -18,11 +18,11 @@ export const ServerList = [
     /**
      * Wingspan's Files API in Production
      */
-    "https://api.wingspan.app/files",
+    "https://api.wingspan.app",
     /**
      * Wingspan's Files API in Staging
      */
-    "https://stagingapi.wingspan.app/files",
+    "https://stagingapi.wingspan.app",
 ] as const;
 
 /**
@@ -58,9 +58,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "1.0.0";
-    sdkVersion = "2.0.0";
-    genVersion = "2.209.0";
-    userAgent = "speakeasy-sdk/typescript 2.0.0 2.209.0 1.0.0 @wingspan/files";
+    sdkVersion = "2.0.1";
+    genVersion = "2.210.6";
+    userAgent = "speakeasy-sdk/typescript 2.0.1 2.210.6 1.0.0 @wingspan/files";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);